decopling ht_search and solr_query_interface tests
