+++ app/solr_query_interface/src/data_analysis/extracting_diference_solr_results.py
