+++ app/solr_query_interface/src/data_analysis/generate_query_results_in_batch.py
@@ -4,7 +4,7 @@ import sys
 from argparse import ArgumentParser
 
 import pandas as pd
-from ht_full_text_searcher import HTFullTextSearcher
+from ht_full_text_search.ht_full_text_searcher import HTFullTextSearcher
 from ht_search.config_search import FULL_TEXT_SOLR_URL
 from ht_search.ht_query.ht_full_text_query import HTFullTextQuery
 from ht_utils.ht_logger import get_ht_logger
