+++ app/solr_query_interface/src/data_analysis/get_list_ids_fromSolr_results.py
