+++ app/solr_query_interface/src/ht_full_text_search/main.py
