+++ app/solr_query_interface/tests/ht_full_text_query_test.py
