+++ app/solr_query_interface/tests/ht_full_text_searcher_test.py
@@ -2,7 +2,7 @@ import os
 
 from ht_search import config_search
 
-from app.solr_query_interface.src.scripts.ht_full_text_searcher import HTFullTextSearcher
+from ht_full_text_search.ht_full_text_searcher import HTFullTextSearcher
 
 
 class TestHTFullTextSearcher:
